+++ Source/Platforms/net_platform.h
@@ -97,14 +97,14 @@ os9err NetInstall(void);
 
 
 
-os9err MyInetAddr( ulong *inetAddr, ulong *dns1Addr,
-                                    ulong *dns2Addr, char* domainName );
+os9err MyInetAddr( uint32_t *inetAddr, uint32_t *dns1Addr,
+                                    uint32_t *dns2Addr, char* domainName );
 /* Returns <inetAddr>, DNS addresses and <domainName> of this host */
 
 
 
 
-OSStatus netReadBlock( ushort pid, net_typ* net, ulong *nBytes );
+OSStatus netReadBlock( ushort pid, net_typ* net, uint32_t *nBytes );
 /* Read <nBytes> of data at <net->transferBuffer> from <net->ep> */
 /* Result is <err> >0 -> number of bytes read                    */
 /*                 =0 -> no data                                 */
@@ -112,7 +112,7 @@ OSStatus netReadBlock( ushort pid, net_typ* net, ulong *nBytes );
 /* <net->closeit> will be set to true in case of EOF             */
 
 
-OSStatus netWriteBlock( ushort pid, net_typ* net, ulong *nBytes );
+OSStatus netWriteBlock( ushort pid, net_typ* net, uint32_t *nBytes );
 /* Write <nBytes> of data of <net->transferBuffer> to <net->ep>  */
 /* Result is <err> >0 -> number of bytes written                 */
 /*                 =0 -> no data                                 */
